Enable to config ContainerGCPeriod
ContainerGCPeriod is the period for performing container garbage collection.
This variable is hard code to 1 minute and can not be configured. But with
init container(s) in Pod, Pod recreation/deletion is slow, because the init
container(s) is(are) deleted by 1 minute interval periodically container
GC task. We make this configurable and defaults to 10 seconds to reduce
Pod recreation/deletion time.